refactor: Improve exception handling in processes (#29)
This change will improve exception handling of processes. 

At the moment, when a process throws an uncaught exception, the kernel will catch and log the exception, stop the offending process and then continue. This approach however might cause the user to overlook possible exceptions in processes and does not give any ability to the user for handling these exception.

This change modifies the kernel implementation and specification such that the `step()` method  (and consequently `run()`) must propagate uncaught exceptions that occur in processes. This allows the caller to control the way exceptions are handled.