refactor: Introduce initial API design for 2.x
This change introduces the new API design that will be introduced in the
2.x versions of the OpenDC Simulator.

This changes focuses on simplifying simulation primitives provided by
the simulator and introduces a new concept of actors based on the model
designed by the Akka Typed project.

For now, the old simulation models have been removed from the branch,
but will be ported back as this branch is being finalized.