bug: Fix React invariant error due to dependency mismatch
This change fixes an issue with React crashing due to an invariant error
that was caused by a dependency mismatch between React and react-konva.

I have included the Yarn lock file in the repository to prevent these
issues from happening in the future and keeping the installation
deterministic. Additionally, I have updated react-konva and konva to the
latest version to be compatible with the targeted React release.

Fixes #63