refactor: Remodel workloads and compute resource
This change remodels our model for workloads and compute resources in
order to support VM/container functionality where multiple workloads run
on a single machine.

In particular, we make the following changes:

- Move the compute-related source code into the `opendc-compute` module.
- Change from application-based model to image-based model, where the
image has exclusive access over the machine, instead of applications
that share the machine. We may model in the future again
applications/operating system, but at the moment, we do not need this
granularity, given that the workload traces can be simulated using VMs.