Add module for simulation-specific code for OpenDC
This change introduces a new module in the OpenDC namespace where
simulation-specific code will be placed. We will move away from the
odcsim naming to prevent getting confused users due to this complex
naming scheme.

Furthermore, this module will live in the org.opendc package instead of
the com.atlarge.opendc to further reduce depth of the packages.