Remove odcsim components from repository
This change removes the odcsim components from the repository as we have
eliminated their use in the OpenDC codebase, by replacing them with the
more generic (Test)CoroutineScope and Clock. From now on, we will only
place modules under the OpenDC namespace and not use odcsim as well to
prevent confusion.