ci: Build pushes to master branch
This change updates the build workflow to trigger on pushes to the
master branch. We previously removed this trigger to reduce the number
of CI builds when merging changes into the master branch.

However, with the integration of Codecov as well as the caching steps in
the workflow, it makes sense to build every push to the master branch.