ci: Update Java Github Actions workflows
This change updates the Github Actions workflows for the Java/Kotlin
part of the project. We now use the gradle-build-action from Gradle to
build the project, which does caching for us.

In addition, we update the Codecov action to version 2, since the old
version is deprecated and will stop working next year.