refactor(workflow): Remove WorkflowSchedulerMode
This change removes the WorkflowSchedulerMode interface in favour of an
integrated timer scheduler approach that batches scheduling cycles over
a user-specified quantum. This quantum can be lowered to a small value
to get the interactive behavior.

There is no replacement for the random behavior, but we believe that
such a policy makes no sense in a real-world scenario.