feat(trace/opendc): Incorporate interference model in trace format
This change updates the OpenDC VM trace format to incorporate the VM
interference model in the trace format itself. This makes sense since
the model is tightly coupled to the actual trace that is being
simulated.

This approach has as benefit that we can directly load the
interference model from the workload trace, without having to resolve
the model seperately (as we did before).