refactor(compute): Expose CPU and system stats via Host interface
This change updates the `Host` interface to directly expose CPU and
system stats to be used by components that interface with the `Host`
interface.

Previously, this would require the user to interact with the OpenTelemetry SDK.
Although that is still possible for more advanced usage cases, users can
use the following methods to easily access common host and guest
statistics.