refactor(web/server): Create standalone OpenDC distribution
This change updates the Quarkus configuration of the OpenDC web server
to serve as a fully standalone distribution that is capable of serving
the web UI, web API, and experiment runner. Such an approach vastly
simplifies local deployments.

For Docker deployments, we create a custom Quarkus profile that uses
PostgreSQL and disables the web UI.