build: Remove unused distribution conventions
This change removes the distribution conventions used by the root
project of OpenDC. Previously, we used this to build a single
distribution for OpenDC containing the experiment uberjars and scripts
to start OpenDC.

However, with the introduction of the Quarkus-based web server, we have
decided to split releases into (potentially) multiple distributions
(e.g., one for the web server, one for just the web runner, etc.).

Furthermore, the implementation of this convention caused issues with
several other Gradle plugins.