refactor(compute/api): Do not suspend in compute API
This change updates the API interface of the OpenDC Compute service to
not suspend execution using Kotlin Coroutines.

The suspending modifiers were introduced in case the ComputeClient would
communicate with the service over a network connection. However, the main
use-case has been together with the ComputeService, where the suspending
modifiers only frustrate the user experience when writing experiments.

Furthermore, with the advent of Project Loom, it is not necessarily a
problem to block the (virtual) thread during network communications.