ci: Migrate to GitHub Container Registry
Docker Inc is sunsetting free team organizations for the Docker registry,
which our organization is one of. Instead, a paid subscription is now required
to maintain the organization.

Given our relatively small usage of the account, it makes more sense to start
publishing the container images on the GitHub Container Registry, since it is
free for open source projects and integrates well with GitHub Actions.

Fixes #141