bug(web): Fix access to machines on lower shelves
This change addresses #137 which reports that machines on shelves lower
than the top shelve cannot be accessed and doing so will cause the UI to
disappear and an error message to be generated.

The issue was caused by using the incorrect logic for selecting the
machine at a certain rack position.

Fixes #137